Fix for Github issue #17497 - Invalid RA diagnostic error: expected 2 arguments, found 1
The issue occurs because RA discards the `mut` qualifier that exists in the code being analyzed,
in the case where a variable is an owned value, not a ref.

E.g. `let mut vec = Vec::new();`

Then, when resolving candidate methods against this variable as a receiver, RA incorrectly
rejects all `&mut self` methods, because it has lost the information that the variable is mutable.

The fix involves preserving the `mut` qualifier information when carrying out method resolution.
At the point that the code fails to unify the receiver's type with a candidate method's expected receiver type,
an additional chalk request is done but this time with the receiver type coerced to &mut.

Includes a unit test that only passes after applying the fixes in this commit.